More command work.
Note to self. Make new bot token.

+++ Discord Groups/Discord Groups - Server/src/com/weeryan17/dgs/listeners/discord/ChatListener.java
@@ -2,6 +2,7 @@ package com.weeryan17.dgs.listeners.discord;
 
 import java.util.Arrays;
 
+import com.vdurmont.emoji.EmojiManager;
 import com.weeryan17.dgs.DiscordGroups;
 
 import sx.blah.discord.api.events.EventSubscriber;
@@ -21,12 +22,16 @@ public class ChatListener {
 		IMessage message = e.getMessage();
 		String text = message.getContent();
 		if(prefix == text.charAt(0)){
-			message.delete();
 			text = text.substring(1);
 			String[] command = text.split(" ");
 			String name = command[0];
 			String[] args = Arrays.copyOfRange(command, 1, command.length);
-			instance.getCommandMannager().dispatchCommand(name, args, e.getChannel());
+			if(instance.getCommandMannager().isCommand(name)){
+				message.delete();
+				instance.getCommandMannager().dispatchCommand(name, args, e.getChannel(), e.getAuthor());
+			} else {
+				message.addReaction(EmojiManager.getForAlias("question"));
+			}
 		}
 	}
 }
